+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 
 * Added support for dark theme mode on Windows (including windows title bar background),
   MacOS and GNU/Linux.
+* Added embbeded Qt-based Python console widget
 * Dataset edit layout: now disabling/enabling "Apply" button depending on widget value changes
 * Code editor: widget minimum size area may now be set using rows and columns size
 * Test launcher: redesigned, added support for dark mode